backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
requests>=2.31
PyJWT>=2.8
gunicorn>=21.2
pytest>=7.4
pytest-cov>=4.1
